[add class average, variance] one-way anova, two-way anova
