+++ statistics/src/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from calculate_average import CalculateAverage
+from calculate_variance import CalculateVariance
+
+if __name__ == '__main__':
+    pass
